Move RunEnergy plugin into StatusOrbs plugin
